@@ -7,6 +7,6 @@ module BookmarkSystem
   ###
   # Version constant definition
   ###
-  VERSION = "0.0.9"
+  VERSION = "0.1.0"
 end
 
